requirements.txt
attrs==25.1.0
black==25.1.0
certifi==2025.1.31
click==8.1.8
Faker==36.1.1
h11==0.14.0
idna==3.10
iniconfig==2.0.0
mypy-extensions==1.0.0
outcome==1.3.0.post0
packaging==24.2
pathspec==0.12.1
platformdirs==4.3.6
pluggy==1.5.0
PySocks==1.7.1
pytest==8.3.4
pytest-faker==2.0.0
python-dotenv==1.0.1
selenium==4.29.0
sniffio==1.3.1
sortedcontainers==2.4.0
trio==0.29.0
trio-websocket==0.12.1
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
websocket-client==1.8.0
wsproto==1.2.0
